refactor: renamed modules to plural
