+++ packages/pure-parse/src/guards/arrays.test.ts
