+++ packages/pure-parse/src/guards/arrays.ts
