+++ packages/pure-parse/src/guards/index.ts
