+++ packages/pure-parse/src/guards/literal.test.ts
