+++ packages/pure-parse/src/guards/literal.ts
