+++ packages/pure-parse/src/guards/object.ts
