+++ packages/pure-parse/src/guards/optional.test.ts
