+++ packages/pure-parse/src/guards/optional.ts
