+++ packages/pure-parse/src/guards/primitives.ts
