+++ packages/pure-parse/src/guards/records.ts
