+++ packages/pure-parse/src/guards/tuple.ts
