+++ packages/pure-parse/src/guards/types.ts
