+++ packages/pure-parse/src/guards/union.ts
