+++ packages/pure-parse/src/index.ts
@@ -1,3 +1,3 @@
 export * from './common'
-export * from './parse'
-export * from './guard'
+export * from './parsers'
+export * from './guards'
