+++ packages/pure-parse/src/parsers/literal.test.ts
