+++ packages/pure-parse/src/parsers/optional.test.ts
