+++ packages/pure-parse/src/parsers/types.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, test } from 'vitest'
-import { array } from './array'
+import { array } from './arrays'
 import { object } from './object'
 import { union } from './union'
 import { parseNumber, parseString } from './primitives'
